Glasspane
So, there is a glasspane and tutorial how add gears
There is a problem with doing the tutorial multiple times without reloading the page
This is a problem due to bootstrap, can be fixed pretty easy but I don't have time right now.
Feel free to fix it if you want.
Otherwise, test it on your computers and see if it works!